telegram-bot/requirements.txt
python-telegram-bot>=21.0
google-api-python-client>=2.0
google-auth>=2.0
google-auth-httplib2>=0.2.0
supabase>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
